Add benchmark tests to travis
