+++ package.json
@@ -27,6 +27,8 @@
     "node": ">= 0.4.0"
   },
   "devDependencies": {
+    "benchmark": "^2.1.2",
+    "chalk": "^1.1.3",
     "grunt": "^0.4.5",
     "grunt-bump": "^0.3.0",
     "grunt-contrib-uglify": "~0.5.0",
@@ -38,6 +40,9 @@
     "tsc": "tsc src/json-patch-duplex.ts --noEmitHelpers && tsc src/json-patch.ts --noEmitHelpers",
     "test": "npm run test-core && npm run test-duplex",
     "test-duplex": "jasmine DUPLEX=yes JASMINE_CONFIG_PATH=test/jasmine.json",
-    "test-core": "jasmine  DUPLEX=no JASMINE_CONFIG_PATH=test/jasmine.json test/spec/jsonPatchTestsSpec.js test/spec/coreSpec.js test/spec/validateSpec.js"
+    "test-core": "jasmine  DUPLEX=no JASMINE_CONFIG_PATH=test/jasmine.json test/spec/jsonPatchTestsSpec.js test/spec/coreSpec.js test/spec/validateSpec.js",
+    "bench": "npm run bench-core && npm run bench-duplex",
+    "bench-core": "node test/spec/coreBenchmark.js",
+    "bench-duplex": "node test/spec/coreBenchmark.js DUPLEX=yes && node test/spec/duplexBenchmark.js"
   }
 }
