+++ test/SpecRunner.html
@@ -27,6 +27,7 @@
   <script src="spec/jsonPatchTestsSpec.js"></script>
   <script src="spec/coreSpec.js"></script>
   <script src="spec/validateSpec.js"></script>
+  <script src="spec/coreBenchmark.js"></script>
 
 
 </head>
