+++ test/SpecRunnerDuplex.html
@@ -29,6 +29,8 @@
   <script src="spec/coreSpec.js"></script>
   <script src="spec/validateSpec.js"></script>
   <script src="spec/duplexSpec.js"></script>
+  <script src="spec/duplexBenchmark.js"></script>
+  <script src="spec/coreBenchmark.js"></script>
 
 </head>
