+++ test/spec/coreSpec.js
@@ -728,92 +728,3 @@ describe("undefined - JS to JSON projection / JSON to JS extension", function() 
         });
     });
 });
-
-// Benchmark performance test
-if (typeof Benchmark !== 'undefined') {
-    var suite = new Benchmark.Suite;
-    suite.add('add operation', function() {
-        obj = {
-            foo: 1,
-            baz: [{
-                qux: 'hello'
-            }]
-        };
-        jsonpatch.apply(obj, [{
-            op: 'add',
-            path: '/bar',
-            value: [1, 2, 3, 4]
-        }]);
-    });
-    suite.add('remove operation', function() {
-        obj = {
-            foo: 1,
-            baz: [{
-                qux: 'hello'
-            }],
-            bar: [1, 2, 3, 4]
-        };
-        jsonpatch.apply(obj, [{
-            op: 'remove',
-            path: '/bar'
-        }]);
-    });
-    suite.add('replace operation', function() {
-        obj = {
-            foo: 1,
-            baz: [{
-                qux: 'hello'
-            }]
-        };
-        jsonpatch.apply(obj, [{
-            op: 'replace',
-            path: '/foo',
-            value: [1, 2, 3, 4]
-        }]);
-    });
-    suite.add('move operation', function() {
-        obj = {
-            foo: 1,
-            baz: [{
-                qux: 'hello'
-            }],
-            bar: [1, 2, 3, 4]
-        };
-        jsonpatch.apply(obj, [{
-            op: 'move',
-            from: '/baz/0',
-            path: '/bar/0'
-        }]);
-    });
-    suite.add('copy operation', function() {
-        obj = {
-            foo: 1,
-            baz: [{
-                qux: 'hello'
-            }],
-            bar: [1, 2, 3, 4]
-        };
-        jsonpatch.apply(obj, [{
-            op: 'copy',
-            from: '/baz/0',
-            path: '/bar/0'
-        }]);
-    });
-    suite.add('test operation', function() {
-        obj = {
-            foo: 1,
-            baz: [{
-                qux: 'hello'
-            }]
-        };
-        jsonpatch.apply(obj, [{
-            op: 'test',
-            path: '/baz',
-            value: [{
-                qux: 'hello'
-            }]
-        }]);
-    });
-
-    benchmarkReporter(suite);
-}
