+++ test/spec/duplexSpec.js
@@ -1493,59 +1493,3 @@ describe("duplex", function() {
         });
     });
 });
-
-// Benchmark performance test
-if (typeof Benchmark !== 'undefined') {
-    suite.add('generate operation', function() {
-        obj = {
-            firstName: "Albert",
-            lastName: "Einstein",
-            phoneNumbers: [{
-                number: "12345"
-            }, {
-                number: "45353"
-            }]
-        };
-        var observer = jsonpatch.observe(obj);
-
-        obj.firstName = "Joachim";
-        obj.lastName = "Wester";
-        obj.phoneNumbers[0].number = "123";
-        obj.phoneNumbers[1].number = "456";
-
-        var patches = jsonpatch.generate(observer);
-        obj2 = {
-            firstName: "Albert",
-            lastName: "Einstein",
-            phoneNumbers: [{
-                number: "12345"
-            }, {
-                number: "45353"
-            }]
-        };
-
-        jsonpatch.apply(obj2, patches);
-    });
-    suite.add('compare operation', function() {
-        var obj = {
-            firstName: "Albert",
-            lastName: "Einstein",
-            phoneNumbers: [{
-                number: "12345"
-            }, {
-                number: "45353"
-            }]
-        };
-        var obj2 = {
-            firstName: "Joachim",
-            lastName: "Wester",
-            mobileNumbers: [{
-                number: "12345"
-            }, {
-                number: "45353"
-            }]
-        };
-
-        var patches = jsonpatch.compare(obj, obj2);
-    });
-}
